+++ src/Constraints.cpp
@@ -1,13 +1,13 @@
 #include <cmath>
 #include "Constraints.hpp"
 
-double Constraints::constrainedSpeed(double curvature) {
-    double maxTurnSpeed = 2 * maxVel / (2 + std::abs(curvature) * trackWidth);
+float Constraints::constrainedSpeed(float curvature) {
+    float maxTurnSpeed = 2 * maxVel / (2 + std::abs(curvature) * trackWidth);
     if (curvature == 0) { return maxTurnSpeed; }
-    double maxSlipSpeed = std::sqrt(frictionCoeff / std::abs(curvature) * 9.81);
+    float maxSlipSpeed = std::sqrt(frictionCoeff / std::abs(curvature) * 9.81);
     return std::min(maxTurnSpeed, maxSlipSpeed);
 }
 
-double Constraints::constrainedAcc(double curvature) { return 2 * maxAcc / (trackWidth * std::abs(curvature) + 2); }
+float Constraints::constrainedAcc(float curvature) { return 2 * maxAcc / (trackWidth * std::abs(curvature) + 2); }
 
-double Constraints::constrainedDec(double curvature) { return 2 * maxDec / (trackWidth * std::abs(curvature) + 2); }
+float Constraints::constrainedDec(float curvature) { return 2 * maxDec / (trackWidth * std::abs(curvature) + 2); }
